+++ person.rb
@@ -1,8 +1,8 @@
 class Person
-    def initialize(name = "Unknown", age, parent_permission = true)
-        @name = name
-        @age = age
-        @parent_permission = parent_permission
+  def initialize(name = "Unknown", age, parent_permission = true)
+    @name = name
+    @age = age
+    @parent_permission = parent_permission
    end
    
    def get_id()
@@ -18,25 +18,24 @@ class Person
     end
 
     def set_id(id)
-        @id = id
+      @id = id
     end
 
     def set_name(name)
-        @name = name
+      @name = name
     end
 
     def set_age(age)
-        @age = age
+      @age = age
     end
     
     private
     def of_age?
-        @age >= 18
+      @age >= 18
     end
 
     public
     def can_use_services?
-        @age >= 18 || @parent_permission == true
+      @age >= 18 || @parent_permission == true
     end
-
 end
